test: extract component fixture to fixtures/custom-resources/, move to resources/custom-resources.test.ts
Addresses Ethan Arrowood's review feedback on #1211: inline SCHEMA_GRAPHQL,
RESOURCES_JS, and CONFIG_YAML strings extracted to real files under
integrationTests/fixtures/custom-resources/. Test moved from
integrationTests/apiTests/ to integrationTests/resources/ per directory
conventions. Deployment unchanged: installAppComponent still receives
name→content maps, now loaded with readFileSync at test startup.
